ecommerce-mongodb/requirements.txt
beanie==1.29.0
fastapi==0.115.12
motor==3.7.0
passlib==1.7.4
pydantic==2.11.3
PyJWT==2.8.0
uvicorn==0.34.1
